refactor: move e2e tests for folder tests/e2e
